move API calls to sqlite worker impl
